MB-60560: Compile ns_server and deps with rebar3
Move to using rebar3 to compile ns_server code. This change is linked
to a corresponding tlm change by Change-Id.

Some notes:

We are moving all of our (ns_server owned/core) applications to a new
apps subdirectory. This is in line with the "umbrella" project
structure type. The applications that we are moving are ale,
ns_babysitter, ns_common, ns_couchdb, and ns_server itself. We are
moving these applications here, and out of deps, because that's the
rebar3 "umbrella" project structure. rebar3, in fact, says that we
should no longer maintain a deps folder, but we must push the rest of
our applications out to a remote host (i.e. github) to pull them in the
normal rebar3 way. We do not want to move all of the deps to apps
(temporarily) because rebar3 compiles deps in prod (default) profile
even when compiling a higher level application in test profile which is
desirable to use as compiling and running these tests would take
additional time. For now we will keep the deps folder and utilise the
rebar3_path_deps plugin to "pull" them from the local copy.

We are moving the ns_server application to apps because rebar.config
options at the highest level (ns_server/rebar.config) are applied to
the rebar.config options of the applications in apps. This means that
we would introduce dependencies on applications that do not make sense
(such as ns_common requiring ale's parse_transform ale_transform). To
solve this issue we move the ns_server application to the apps
directory and provide a generic rebar.config at the highest level.

Due to the move of many applications from deps to apps, the review is
somewhat difficult in gerrit. /Most/ files are "Renamed" but some show
up as a removal and a deletion.

Due to the move of applications from deps to apps and the use of rebar3
to compile the codebase, we no longer need CMake driven compilation for
our applications or dependencies. Dependencies can now all be listed in
rebar.config, and only a top level CMakeLists.txt is needed (for our
erlang code at least, we still have some go code in deps). As a result,
we must now specify INSTALL directives for dependencies in the top level
CMakeLists.txt.

Common options for our apps (apps/*) are now specified in the top level
rebar3 file only.

/test/fake_helpers has been unpacked into /test. Rebar3 will build it
if we tell it to (via erl_opts/extra_src_dirs) but it will unpack
everything to that directory which is a bit rubbish and unnecessary.
It's easier to move the two files back under /test.

Change-Id: I23ddb91d66d5ac73bc972fef6c0c906841b71be5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202086
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>