MB-60560: Move cut to new ns_common dep
In a future change to use rebar3 to compile ns_server et. al. we
must declare a dependency for ns_couchdb containing cut.hrl/.erl.
It does not make much sense to recompile the entirety of ns_server
to compile ns_couchdb, and it was not trivial to make rebar3 compile
ns_server as a dependency of ns_couchdb; ns_server would not properly
recompile it's dependencies which was not observed when compiling
ns_server standalone. The issues, I believe, were due to some
combination of the need to symlink dirs in _checkouts and rebar.lock,
but this solution feels much more practical and desirable than fighting
against rebar.

Note that in this change many instances of `-include` have been changed,
primarily for `cut.hrl`. This is because we are now finding `cut.hrl`
from the ns_common library and we'd otherwise have to override the `-i`
variable in the cut rebar.config which:

a) is very questionably correct
b) requires more work to set the path correctly (see how we do this in
   the ns_server rebar.config at the moment)

Change-Id: I5f64dd07d6418b0b6dd7e52949267aece574e059
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202131
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>