MB-59880: Make failover async to auto_failover module
The auto_failover module currently blocks when it attempts to fail over
a node until the orchestrator call returns. This can take a not
insubstantial amount of time when compared to the configured
auto-failover timeout value. For example a failover could take 2 seconds
to process but with an auto-failover timeout value of 1 second  we could
see a maximum of around 3 seconds between a node going down and
auto_failover triggering a failover if the node goes down at the point
at which auto_failover begins failing over a different node.

This code change aims to address this issue by performing the failover
on a background process and informing auto_failover when it is done.
Auto_failover should not attempt to fail over multiple nodes at once as
that would result in the later failovers failing if one is already in
progress. Instead, auto_failover should just retry the action next tick.

Change-Id: If727e3efa7890ea079133872a9a591c20c14ecff
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204763
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>