MB-59676 - add four new CBAS stats, mark four existing stats as deprecated
See the bug ticket for screen shots of the validation.

This was tricky - there were a number of new stats for 7.6 which had
been added in a way that they would appear in mixed 7.2 / 7.6
clusters, yet not be usable. This change did the following:

- when upgrading to 7.6, change the deprecated stats to the new ones

- in a mixed cluster, only show the regular 7.0 stats

- when the cluster is fully 7.6, add the new stats, and mark the
  deprecated stats as deprecated.

- also change the default dashboard stats, so that new "All Services"
  dashboards use the non-deprecated stats in 7.6.

Change-Id: I1c504230cec874e2559dc917f3f652cb26b725f0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/203396
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>