MB-60175: Don't reset canAbortRebalance when ...
... disabling auto-failover. For some reason, the code was written to
disable canAbortRebalance whenever we disable auto-failover and that
value will stick even if it's later enabled again. This would cause the
value to flip-flop without the user ever doing anything to cause it.
This is very confusing so this change should make the behavior more
consistent.

Change-Id: If11430db36bd4bb77621a3ceff2d9d8e0e1a4511
Reviewed-on: https://review.couchbase.org/c/ns_server/+/203560
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>