MB-60311: Cbcollect not correctly cleaning /tmp
It turns out we are not correctly deleting the temporary directory in
use, unless the program is killed while running. Because of the redesign
of cbcollect, we don't store nearly as much in that /tmp directory, so
it looks like we only leave whatever files are created in "artifacts".
Artifacts are files created by another task running and are the only
files that can't be directly routed to the output zipfile in a streaming
fashion. So that explains why the reporter of this bug noticed only
kv_trace.json and systemd_journal.gz in the /tmp directory, because
those are the only files ever placed there. So the fix is to make sure
we call finalize() instead of close() at the end of using the
TaskRunner.

Also made sure to allow errors to bubble up if we do fail to delete any
files in the temporary directory.

Change-Id: I8480030976fe81175f8d14afa005a553604c3ea3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/203561
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>