MB-57610[BP]: Rebalance progress metrics
BP of https://review.couchbase.org/c/ns_server/+/203864

At each scrape, there will be one metric for each service, labelled as
"stage". Once the rebalance starts, the progress will be set to 0 for
all stages. Once the rebalance of that service completes, the progress
for that stage will be set to 1. If the rebalance terminates early, or
is stopped, the progress will be left at the last value reached, until
the next rebalance starts and it gets set back to 0.

Change-Id: I23a6a04846e14f85a4c1f83ee0879c61103e7b91
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204481
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>