MB-57610: Rebalance progress metrics
At each scrape, there will be one metric for each service, labelled as
"stage". Once the rebalance starts, the progress will be set to 0 for
all stages. Once the rebalance of that service completes, the progress
for that stage will be set to 1. If the rebalance terminates early, or
is stopped, the progress will be left at the last value reached, until
the next rebalance starts and it gets set back to 0.

Change-Id: I6c6f9e3ffcd10b10475751165bc97cd82b957391
Reviewed-on: https://review.couchbase.org/c/ns_server/+/203864
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>