MB-57610: Configurable ns_server_stats gauge expiration
For some metrics, the default expiration time of 3mins may not be
appropriate. We should be able to specify the expiration, including
infinity for no expiration.
However, we should only use infinity when updates to the metric could
be arbitrarily far apart, and the metric is expected to be valid
indefinitely. If the metric is associated with an entity which could
disappear (for instance bucket deletion, or potentially in future, the
removal of a service), then the gauge should get cleaned up at that
point.

Change-Id: I8c1bebc44eefc8954aa54e657ea0dc8cce14a1f3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204044
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>