MB-57610[BP]: Configurable ns_server_stats gauge expiration
BP of https://review.couchbase.org/c/ns_server/+/204044

For some metrics, the default expiration time of 3mins may not be
appropriate. We should be able to specify the expiration, including
infinity for no expiration.
However, we should only use infinity when updates to the metric could
be arbitrarily far apart, and the metric is expected to be valid
indefinitely. If the metric is associated with an entity which could
disappear (for instance bucket deletion, or potentially in future, the
removal of a service), then the gauge should get cleaned up at that
point.

Change-Id: I6c0e492abba1224e95e656f0c3f94c3fe7d9d5ef
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204480
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>