MB-60479: ignore FileNotFoundError cbcollect rmdir
When deleting tmp and other directories using shutil.rmtree we use a
callback to log any errors that occur. One situation where you don't
need to do this is when we get a FileNotFoundException because not
finding the file is the same outcome as deleting would have had so we
don't care. Still logs all other classes of errors that could occur.

Change-Id: Ib6de0db4d68783d8aa5514ef3fa0f699f44b5924
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204142
Well-Formed: Restriction Checker
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>