MB-60429: auth audit bad domain rejected->unknown
Change the domain of failed auth events from "rejected" to "unknown" s/t
it can be decoded by memcached properly. It was being dropped previously
because the field of domain was considered required and we were sending
"rejected". After discussion with kv team it was decided that "unknown"
was a better replacement for this use case and that they would begin to
accept unknown as the domain.

Additionally created a cluster_test for audit log testing, and
specifically testing this change from rejected->unknown.

Change-Id: Iddfa6f70f47460f5001a9651c4d394d4ad079704
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204286
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>