MB-60554: Use rebar config script instead of cmake configure file
Rebar config scripts were written for this sort of thing and can be
specified in erlang. This will be used in the future to allow us to
build in different directories.

We chose to avoid making the build crash if the expected environment
variables are not set because it would prevent us from running manual
invocations of `rebar` such as `rebar tree` without the appropriate
environment. To pass the correct environment, we would have had to
create make targets for these commands.

Change-Id: I03d9a0bd75044c40e5190b7861d28ac382bfa7da
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204592
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>