MB-60429: audit properly convert unknown domains
There was an error introduced recently when we began accepting "unknown"
as a valid domain to filter user audit logs which would accept unknown +
any characters after, but would get saved as just "unknown". This isn't
the end of the world but it is confusing so this patch aims to address
this by more carefully converting the entered domain into "unknown"
while still properly handling incorrect/invalid domains.

Change-Id: Ia5f14fc426643a5917d81c4dea8096e69914e4e3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/204618
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>