MB-60621 Stat description missed in merge conflict
When merging forward https://review.couchbase.org/c/ns_server/+/205155
from neo to 7.6 the description for the new stat was lost. It likely
occurred due to human confusion to the merge conflicts arising from
the alphabetizing of the stat names.

This change adds the missing stat description.

Change-Id: I93f0f450090ae88a1cf700ec154e6d53afa5a6b5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/205268
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>