MB-60763 Handle _system collection's requirement
Added a comment which lists the requirement for collections within the
_system scope to begin with "_". This is to prevent future issues from
arising. This requirement was listed in the comments for:

https: //docs.google.com/document/d/
        112dmYu_gyUfs5eKwopH7s1vrZBfeXNlKc8mlU1GGnZQ/edit?usp=sharing

Also, add a unit test to verify adding a collection to the _system scope
is not allowed as well as verifying predefined collections in _system
scope begin with "_".

Change-Id: I78a240ea6df32e2773a9abf67191737b9bdc61e8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/205506
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>