MB-60652 RestEjectTest waits for bucket on all nodes
When doing a creation of the bucket used in the test we now wait for it
to be ready on all nodes. This prevents the case where we want to do a
failover but the bucket is only ready on the node being failed over.

Change-Id: Ifacbb5141a3c2d755c8754330f82746c7190ef12
Reviewed-on: https://review.couchbase.org/c/ns_server/+/206055
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>