MB-58882: Ensure that rebalance alert is disabled in UI
In 7.6, we added kv rebalance alert which was 'disabled' by default,
and not added to the UI. The intention was that we weren't confident
that the alert was reliable enough to enable by default, but we wanted
to be able to enable it manually if desired.

The way this was implemented involved setting the default threshold to
undefined, but this means that now that we've added them to the UI,
they don't actually show as disabled. To do this properly, we need to
use the config keys to actually disable the alert in the way that the
UI properly displays.

This patch sets a reasonable default for the threshold, and ensures
that the pop-up and email are properly disabled by default.
This involves upgrading the config to disable the alert, but if the
alert has already been enabled by manually setting a threshold, we
leave it as it is.

Also adds a missing short description for the alert, which is used in
the email subject, when the alert is enabled.

Change-Id: Ibea1735ac97fcb19049e83e51f1d4a84b3f79ff7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/206278
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>