MB-60924/60107 Set fail_if_no_peer_cert to false
When requiring client certificate is set to enable we want to let the
user decide whether or not to use a certificate. A change was made to
erlang 26 which requires setting fail_if_no_peer_cert to false in this
case; otherwise a certificate is required.

Change-Id: Ia62057d759846e3a5615a0e5fac8b58e773eeae6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/206296
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>