MB-60642: [cluster_test] incorrect Host header
Cluster test to simulate the incorrect Host header issue, as well as
test the mitigation that just involves redirecting to a relative url.
This allows us to sidestep the issue of the Host header and worrying
about the redirect being directed to another host/protocol/etc.

Change-Id: I3ac72a25a910e202b5fc2201f9c426173166612c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207118
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>