MB-68545[BP]: Fix yaml:encode test
Note: this BP depends on Erlang/OTP 26, but since it fixes a commented
out test, it is fine to upgrade to OTP 26 before merging this.

Original commit message below:
-----------------------------

The maps:to_list is non-deterministic wrt the order in the list of the
items. In order to get deterministic ordering we use an ordered maps
iterator.

Change-Id: Ia79245e5eddcedf158977885fabfb3b771602401
Reviewed-on: https://review.couchbase.org/c/ns_server/+/233914
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>