MB-61010: [cluster_tests] Workaround for "EOF occurred...
... in violation of protocol" problem

Sometimes TLS connection fails with reason "EOF occurred in
violation of protocol" while ALERT_CERTIFICATE_REQUIRED is
expected.
It is not clear why it happens at this moment. Since it is not
a critical failure, adding a workaround that just retries the
connection attempt..

Change-Id: Id32920d2d0c6de5e4a116b61de5d8f0f014635e8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207725
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>