MB-57610: Correct "added" version for cm_rebalance_progress
The metric was originally added in the master branch, before being
backported to the trinity branch. That meant that it was added in 7.6.2
but the "added" version field got the incorrect value of 8.0.0.

This change sets the "added" field to 7.6.2 for cm_rebalance_progress
in etc/metrics_metadata.json

Change-Id: I6513f04a2b75d8b2bf38a32fbe2473e65e1c1f1b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207978
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>