MB-58691: Config/Node remap script
A script to facilitate remapping a node by rewriting the configuration
that ns_server uses while the node is offline. Whilst the primary
purpose of the script is to remap nodes, it can, in theory, be used
to remap arbitrary configuration.

Change-Id: Ia404f45835caa2e39d8769501e6fed80c4239f22
Reviewed-on: https://review.couchbase.org/c/ns_server/+/208046
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>