MB-61072 Give read-only admins `backup!read` perms
This is to allow the backup service UI to be accessed so they can check
task history. The permission will allow viewing repositories, their
tasks and metadata about the tasks (e.g. a summary of how many documents
were backed up) but not viewing actual documents in backups or perform
any actions like merge or restore.

This commit also changes `backup!all` to `backup!read` in
the list of permissions we always observe, as we now use the latter to
show/hide the backup tab in the UI.

Change-Id: If388c78df584c159ea65f149621f3f7a7c725fc5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/208679
Tested-by: Matt Hall <matt.hall@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Peter Searby <peter.searby@couchbase.com>