MB-61292: [encryption_service] Load config refactoring
Before this change we assumed that config contains only Secret
Manager (SM) configuration, so when copy_secrets retuned "same"
we treated that as "config has not changed".
This approach leads to the following problem:
If we add new settings that are unrelated to SM, it will be
impossible to load a config that has changes only for new settings.

This change addresses the problem described above by continuing
config reload if we copy_secret returns "same". It also makes sure
that we don't try to remove "old secrets" after reload is complete
(because no copying was done before that).

Change-Id: I934780bc2107df5750302e6f6f802fa698128bb0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209123
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>