MB-61669: Persist state on autofailover completion
Failover is now async to auto_failover module due to MB-59880. The
autofailover count kept in the state is always read from ns_config on
each tick in autofailover module and then the state is persisted at
the end. This will cause the failover count be lost/overwritten. This
change also persists the state at the end of auto_failover completion
so that it is not lost.

Change-Id: Iae58d92dd174d311d5ef6eb3a5173907d1774ef1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209166
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>