MB-59113/MB-61201: Disable indexing guardrails
Indexing guardrails are no longer targeted for an upcoming release,
(MB-61201) so they should be disabled.

Ideally we would just revert the patches, but since there's quite a
large number of them, and they're mixed in with some code/test
improvements, it's more work than it's worth to revert them.

Instead, we are just disabling the ns_config configuration and removing
the metakv configuration, so no checks will be performed in ns_server,
and no extra config keys will be sent to indexing (which currently
cause errors in the logs, since they are unhandled).
The REST API configuration is also disabled, via the config profile.

Additionally, because the index diskUsage API was originally added as
extra thresholds in the diskUsage configuration, this needed to be
moved under the index guardrails configuration section. This allows
us to easily disable the index disk usage guardrail, including the API.

We previously disabled them in the trinity branch, after they were no
longer desired for 7.6.2, but re-enabled them in master so that
indexing could continue development.
Since they are no longer targeted for morpheus, we should disable these
guardrails in master again. We can re-enable them when indexing is
ready for the settings.

Change-Id: Ice336ef3ac8f55c8d99594c54f3e1e9e51b4fcf2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213465
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>