Revert "MB-61201 Disable indexing guardrails"
This reverts commit 4bce9cd65ae2ea79bdec6f3f84198a32adc2cd00.

Indexing guardrails were originally merged in trinity, targeted for
7.6.2, but they are now targeted for 7.6.3, so they were disabled in
the trinity (7.6.2).

They will likely eventually be re-enabled in 7.6.3, but for now they
should be re-enabled in master, so allow indexing to test against them,
until the trinity branch becomes 7.6.3 and we can backport the change
(if indexing guardrails are ready for 7.6.3).

Change-Id: I731197097fb2edca121b6a47ebe8ada83e760d94
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209685
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>