cluster_tests: Add Balanced requirement
If the cluster is not balanced then it may behave in unexpected ways in
certain tests, since they may assume that for example when creating a
bucket, it will be placed on all nodes.

To avoid this issue, a Balanced requirement should be added to the
requirements of any testset that requires it. This requirement will
be needed whenever it is assumed that the cluster is balanced at the
start of a testset.

An alternative approach considered was to simply add a balanced check
in the smog check, to fail a test if it ever leaves the cluster in an
unbalanced state. The reason this approach was abandoned was that it
adds an additional burden to every testset to need to leave the cluster
in a balanced state. By having the testset opt in to balanced clusters,
we can ensure that any testset that needs a balanced cluster gets one,
without every testset needing to do extra work at the end potentially
unnecessarily.

Now that we don't always rebalance multi-node clusters when creating
them, it's possible for there to be a race between cluster_connect
returning and the cluster actually being able handle /pools/default
(see MB-62153). To account for this, we now poll for /pools/default to
return 200 or 404, when validating that the cluster is up in
cluster.get_cluster.

This covers the following testsets as of this commit:
- BucketMigrationTest
- MultiNodeBucketTestSet
- HardResetTests
- NodeRemovalStreamingChunked
- NodeRemapTests/num_nodes=2,services=all
- RestEjectTest

Change-Id: I9b3e89d020e9aaa3558ec23991d6242a1a74a4dc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209838
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>