MB-61810: Add cm_is_balanced stat
The cm_is_balanced stat is a gauge returning 0 (false) or 1 (true) on
whether or not the cluster is balanced. It is the same value as returned
by "isBalanced" from terseClusterInfo.

The stat is populated into the ns_server_stats ets table by a separate
proess so as to not block other work by what could be a time consuming
operation. As it is a fairly heavyweight stat it is populated less
frequently (every 30 seconds) and only reported by the orchestrator
node.

Change-Id: I60714731818005b58b7c38ea0822c96739bc561b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210221
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>