MB-62013: Perform initial wildcard filter for logs dir
If there are a large number of extra files in the logs dir (e.g. if the
audit log path is set to the logs dir), then we currently can take a
long time processing these, since we go through them all one by one.
This can add a significant amount of time to startup, particularly
because we have to perform this processing for each logger we have.

This can be significantly improved by performing a wildcard filter
prior to iterating over the files.
This gave a 20-30X speedup on a dir of 600,000 files in my local
testing.

Change-Id: I97aff9b0189fb8695fdf372ac2101663c7baf27b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210403
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>