MB-62024: Don't set failpartialwarmup=false
This causes KV to enable traffic to value eviction buckets that have not
loaded all data. This is desirable because we would otherwise enable
traffic to buckets that have not loaded all data, and this would present
as dataloss. This has no impact on full eviction buckets.

Change-Id: I7c2349d688fe144090b778d0367c0ea736ceef14
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210700
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>