MB-63169: Add support for multiple DCP connections between nodes
Add support for multiple DCP connections between nodes. Rather than
spin up a single DCP connection between nodes we can now spin up a
configurable number. vBucket streams are assigned by modulus to a
connection.

This change gives improved replication throughput which increases
rebalance performance and can improve workloads such as durable writes
and bulk loads when disks are sufficiently quick.

Some preliminary perf results can be found here -
https://docs.google.com/spreadsheets/d/1hfNvxw-3FPnoip0N0kuO_jCLdIDZIzPGDM9W9l-UTKI/edit?gid=1991683175#gid=1991683175

The number of connections is tracked on a per-Bucket basis. This:

1) Gives us the flexibility to change this for individual buckets,
allowing us to, for example, achieve better durable write performance on
a particular bucket

2) Easily orchestrate the changing of the connections through the
janitor. We change the connection count via the janitor to avoid races
between rebalance and the connection count changing. We avoid different
nodes having different values by using the bucket config supplied from
the orchestrator node. Finally, this allows us to not worry about
compat mode changes in the janitor, as the APIs will not change.

Of particular note is the connection name handling. Ideally we would
name all connections with some number. Unfortunately, we rely on the
connection naming on both the producer and the consumer nodes, meaning
that we have to consider versioning when changing the format. Adding a
compat mode check would not be ideal, we calculate connection names
on wait_for_data_move loop iterations which could be every 100ms. The
next thing we could do is cache a compat mode or name somewhere. This
could be in the dcp_replicator, but this exists on the consumer node
and would need a network hop/rpc. We would need to cache this somewhere
on the producer which is not particularly intuitive. Finally, to
change the connection name handling like so, we would need to drop any
connections from before the compat mode change on compat mode change,
which is not ideal. It feels far simpler in this case to not supply
a number for the first connection such that the naming on pre and post
morpheus clusters remain the same.

In a future change we will support APIs to change the number of
connections, once the cluster has been ugpraded to morpheus.

Change-Id: I299f3898a15d77cc13884776715cb070aff076ea
Reviewed-on: https://review.couchbase.org/c/ns_server/+/211590
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>