MB-61292: Perform secret modifications on master
... in order to avoid tricky races like the one described below.

During rotation for KEK we generate and encrypt new KEK outside of
chronicle transaction (because we want the transaction to be short).
If we allow parallel secrets modifications, it is possible that
secret  encrypts the newly KEK changes after KEK encryption but
before we add it to chronicle.
Running all modifications in a single process helps us avoid this
kind of races.

Change-Id: I219f142a1e7bc7c6c2952e3752d2b1fc217608e7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/209738
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>