MB-62177: Don't initialise guardrail statuses empty
If a node goes down and comes back up again, the resource_statuses in
ns_config may not be empty. In that case, if the first check sees no
statuses then it wouldn't update ns_config, leaving the monitor in
a state contradicting the ns_config key.

To avoid this, the statuses should initially be fetched from ns_config,
ensuring that ns_config stays in sync with the gen_server state (since
this gen_server is the only process to modify this key).

Change-Id: Iaf99a53ddab10da1162ad3aa6166517c6f36b36f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210909
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>