MB-61292: Check permissions based on secret usage list
In order to make it possible for bucket admin to enable encryption
for the bucket that they manage, bucket admin should be able to
create secrets. But we can't let them create/modify any secret.
Bucket admin should be able to only create/manage a secret that is
to be used for encryption of the bucket that that admin manages.

Change-Id: Ie54bf32f056a8b9041ad52a4a5343f0edb9a36d3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210834
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>