MB-62419: Enable external auth when SAML is enabled
Since SAML users can make requests to memcached via the UI on the docs
page, memcached needs to be able to lookup those users, which means
that external auth must be enabled.

The main alternatives which were considered were changing memcached to
ignore external users when external auth is disabled, and not including
the user's identity when we impersonate external users while external
auth is disabled.

The former solution was rejected by the memcached team as it opens up a
hole in the privilege checking, which could cause unintended behaviour,
and possibly introduce a security vulnerability.

The latter solution was rejected since it means that the user's request
may not be audited fully.

Instead, by simply using external auth, we minimise the creation of
edge-cases that may be poorly handled.

Change-Id: I630d4d634f0ab50d5bdc7869d4c534a02d9acca7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/211650
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker