MB-62174: Add comments explaining disk usage guardrail checks
The linked ticket was filed because it seemed like the existing
behaviour was undesirable, and the intention was to change it.
Upon additional investigaton and further thought, it was realised that
there is a valid justification for the behaviour, that just wasn't
documented anywhere.

To avoid confusion in future, this behaviour should be properly
explained in comments in the code.

Change-Id: Iefb5fc7766b2836fbd4cf933cc38e17a5a758667
Reviewed-on: https://review.couchbase.org/c/ns_server/+/212107
Well-Formed: Restriction Checker
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>