MB-62714: [cbcollect] don't attempt to redact stats
We were accidentally attempting to redact data in the binary stats data.
This wasn't causing any problems because it was never matching, but it
was still incorrectly attempting to redact a lot of data it didn't need
to. Also rename buffer variable because of pylint warnings.

Change-Id: I5ff607c38446a156585eb34c4866fabc03b27589
Reviewed-on: https://review.couchbase.org/c/ns_server/+/212689
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>