MB-62714: [BP] don't attempt to redact stats
We were accidentally attempting to redact data in the binary stats data.
This wasn't causing any problems because it was never matching, but it
was still incorrectly attempting to redact a lot of data it didn't need
to. Also rename buffer variable because of pylint warnings.

Change-Id: I34cbbc40884a0acb19026f798e2da12d2704eb23
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213760
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>