MB-62921: Ensure unit test catches broken offline upgrade
With the original logic, when ns_config_default:get_current_version/0
is updated to a new version, it will automatically mean that offline
upgrade from the previous last minor version would now upgrade to the
new current version, meaning that it skips the previous minor version.

To give an example for clarity, suppose we had versions 7.7 and 8.0,
and we now add 8.1, where the config upgrade path should be
7.7 -> 8.0 -> 8.1
If when adding 8.1, we forgot to update
ns_config_default:upgrade_config/1, we would have the 7.7 upgrade
clause now upgrade the config_version to 8.1, rather than 8.0.
We would also no longer be able to upgrade from 8.0 to 8.1.

We have a unit test which should catch this, but it doesn't because it
simply tests that performing the upgrade doesn't fail, rather than
somehow knowing what versions the upgrade should go through.

Rather than trying to make a more complicated test which would somehow
catch this, we can just update the function itself to fail when we
forget to update it alongside get_current_version/0.

We can do this by always setting config_version explicitly, which means
that when we forget to add the new upgrade clause, it will give a
usupported version error, causing the unit test to fail, as desired.

To give an example for clarity, if we forgot to add the clause for
8.0 -> 8.1 (same as the earlier example), then once the {7,7} clause is
reached, it will set config_version to {8,0}, which wouldn't match the
CurrentVersion clause, so we'd get the desired unsupported version
error.

Note, this is not intended to replace other upgrade testing, which we
should still always perform. This is just intended to catch one case
automatically for us, since it's an easy case to catch (and one which
has been missed initially a couple times now).

Change-Id: I756fffde7ec3b5799c6c02cb94f677c8421fab30
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213424
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>