MB-61320: Make /reloadCertificate no-op for unchanged certs
If the node/client certs, found in "inbox", have not changed from what
is current loaded then /reloadCertificate should simply return. The reason
is the reloading of the certificates is disruptive to the node as it
restarts the TLS distribution and the HTTPs server.

Change-Id: I1f8ffce573f59e6f1ef195f18e159d57863da981
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213961
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>