MB-62777: Make magma the default storage backend
If storageBackend is not specified then it now defaults to "magma". The
default number of vbuckets for magma is 128. At bucket creation time for
a magma bucket the numVBuckets can be specified as either 128 or 1024.
For couchstore buckets numVBuckets can only be 1024 (which is the
default if not specified). Same for ephemeral buckets.

The ramQuota minimum for magma buckets is now 100MB for 128 vbuckets and
1GB for 1024 vbuckets.

Change-Id: I06eacc23147c9994c975abc7756bf43a1c0136e5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/214705
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>