MB-63755: Aggregate activity on orchestrator
In order to have a consistent view of activity in the UI, we need a
single source of truth for activity. Storing this in
ns_config/chronicle would be overkill, and having each node storing it
in replicated_dets would likely mean regular conflicts needing to be
resolved when multiple nodes update the activity for the same user at
the same time. Instead, the orchestrator will aggregate the values,
which it alone distributes to the rest of the cluster via
replicated_dets.

Change-Id: Ia9f294e5abb0f04cdd9002c4859ed8d163fc92d8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/214839
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>