MB-63213: Remove incorrect task __init__
Incorrectly initializes the superclass which causes there to be hidden
behavior where the output of the literal task is interpolated directly
into the header of the task. The easisest solution is removing the
super().__init__() call and accepting a divergence in the class
hierarchy as there is in the functionality. This change was likely an
oversight from the start and shouldn't have been included in the initial
change: a31efb9e389c99e2068a2039ccb3131a0ddc077a

Change-Id: I9b27485c1504d83b36102ffe4b2b2f76bec1447e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/214618
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>