MB-31823: Temporary passwords
Temporary passwords are those that require changing before the user
can perform any other action. This is implemented with a password
expiration time in kv (for future extension to handle expiration of
passwords based on age).

A user can be given (or created with) a single use password with:
PUT /settings/rbac/users/<domain>/<user>
-d temporaryPassword=true
(alongside all existing user properties, which should be repeated to
avoid them being overriden)

This additional field will only be permitted once the whole cluster is
upgraded to morpheus.

When a request comes in from a user with a temporary password, it will
return status 403 and "Password expired", unless the endpoint does not
require a permission check. This means that a user with an expired
password can reach the following endpoints:
- /saml/deauth
- /versions
- /whoami
- /pools/default/trustedCAs - Will only provide certs, nothing else
- /settings/rbac/profiles/@self
- /settings/secrets - Permission check in endpoint handler
- /settings/secrets/<secret id>
- /uilogout
- /controller/changePassword
- /pools/default/checkPermissions - won't have any permissions
- /_cbauth
- /logClientError

The user will also technically have access to endpoints which simply
proxy requests but these are not included above as there are a large
number. These proxy requests will still have a permission check
performed elsewhere, at which point the "Password expired" error would
be returned.

Change-Id: Ib16a52c5a820cd95b5063bbc65f9e238b5d7d973
Reviewed-on: https://review.couchbase.org/c/ns_server/+/215571
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>