MB-63422: Return error if rebalance is not allowed
Prior to this change the orchestrator's rebalance_allowed function would
throw/return "ok" even if the rebalance wasn't allowed. The result of
this was rebalance:start (menelaus_web_cluster.erl) would interpret this
"ok" as a pre-7.6 response.

We now throw/return an error and handle it in the reply to the
requestor.

Change-Id: I01fe37c57503721c2619e43e734b1f3e9510da11
Reviewed-on: https://review.couchbase.org/c/ns_server/+/215659
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>