MB-62604: Pass groups, roles and expiration in cb-on-behalf-extras.
UI session ID is local. If a cb-on-behalf-of request is forwarded
to another node, the forwarded UI session doesn't contain the info
needed to determine AuthnRes. Pass extra_groups, extra_roles and
expiration contained in AuthnRes.

cb-on-behalf-extras contain "session=<..>;groups=<..>;roles=<..>;
expiry=<time> where groups are comma separated and so are roles.
<time> is the time in gregorian seconds.

cbauth change:
https://review.couchbase.org/c/cbauth/+/216211

Backport tests and modify them so that pluggable UI request
is *likely* forwarded to the other node. The other node must decode
the forwarded cb-on-behalf-extras correctly for the test to pass.

Change-Id: I281a8ae0db9c9a07b137ee8c9bdf7f6b9f182329
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216210
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Neelima Premsankar <neelima.premsankar@couchbase.com>