MB-63511: Correct pattern match on #{}
Pattern matching #{} matches a variable to a map TYPE, rather than an
empty map. Instead we should check the size of the map (and type
elsewhere if required).

This caused us not to rewrite the ip files which in turn caused us to
not bring up cluster nodes after remap.

Change-Id: Ie9a597c981bbbf7310f7aab4bc58575ba945d346
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216242
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>