MB-60175/MB-50476: Remove redundant can_abort_rebalance test
The test is trying to verify that the setting is retained when disabling
auto-failover, at least, that's what the commit message implies. The
test doesn't actually appear to do this, and this behaviour should be
covered by the auto_failover_test cluster_test suite.

Change-Id: Id6420e4e8ccf058111705881c9330bd08fa554f6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216243
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>