MB-63586: Remove other LUCKY13 ciphers
There were a few ciphers ending up in our valid list of ciphers for the
menelaus_web tls server. Specifically "AES256-SHA" and "AES128-SHA"
which are not the ones usually listed when you look into LUCKY13 but
they are also CBC block ciphers so it makes sense to remove them.

I also included all of the LUCKY13 ciphers to be removed, even though
they aren't present in the list we remove them from. This is done for
completeness and to guard against this initial list changing in some
way.

Change-Id: I14e8d159a1662f6f13a53350c3b1bf58a6238656
Reviewed-on: https://review.couchbase.org/c/ns_server/+/216610
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>