MB-61292: Call GC for DEKs before doing DEK cleanup
GC is relatively cheap. We ask "users" if some dek is still in use,
and if not, we remove it. At the same time, DEK cleanup can be
expensive. It can involve things like compaction. For this reason,
it seems reasonoble to always try doing GC first, and only if it
doesn't help, call cleanup.

Change-Id: I6933d4df05833284d4348e8558f5cd3fc5a9d630
Reviewed-on: https://review.couchbase.org/c/ns_server/+/217065
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>