MB-62214: XDCR Conflict Logging UI
Adds a new section in XDCR replication page similar to
the Replication Mappings section.
The user can set target locations for logging the replication conflicts.
Locations are defined by a bucket, a scope and/or a collection.

Adds a new property in Bucket Settings dialog: Enable Cross Cluster Versioning.
The new UI section is enabled only if the source bucket of
the replication has this property enabled.
The property is also displayed in bucket details section from the
buckets list.

Changes in Replication Details area from replications list:
* a new expandable section called "Conflict Logging Rules"
displaying the conflict rules in a similar way with "Mapping Rules".
* a new tag displayed near the Source Bucket table cell
visible when the conflict log option is enabled for the replication;
it looks and behaves similar to "Filter" tag.

Change-Id: Ib2537e54d67327a99b3ace8a7ec2fcc6333ba94f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/217332
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>